pyproject.toml
[tool.poetry]
name = "helena-haffner-epic-events-p12-10012025"
version = "0.1.0"
description = ""
authors = ["Featherine <helaena.haffner@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
peewee = "^3.17.8"
bcrypt = "^4.2.1"
pyjwt = "^2.10.1"
python-dotenv = "^1.0.1"
pillow = "^11.1.0"
pytest = "^8.3.5"
coverage = "^7.8.0"
sentry-sdk = "^2.28.0"
flake8 = "^7.2.0"
flake8-html = "^0.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
    ".",
]
